refactor: GeminiTodoAssistant  gemini_provider에서 분리
